refactor(ui/bottom-navigation): optimize code
